requirements.txt
# CivicFix AI Verification Service - Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
imagehash==4.3.1
numpy==1.26.2

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
scikit-learn==1.3.2
scipy==1.11.4

# EXIF/Metadata
piexif==1.1.3
exifread==3.0.0

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Image Analysis
google-cloud-vision==3.5.0  # Optional: Google Vision API
imagededup==0.3.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
